requirements.txt
transformers
torch
scikit-learn
pandas
seaborn
matplotlib
datasets
google-generativeai
python-dotenv